+++ git/git.go
@@ -5,23 +5,16 @@ package git
 
 import (
 	"context"
-	"errors"
-	"fmt"
 	"net/http"
 	"path/filepath"
 	"strings"
 	"time"
 
-	gobilly "github.com/go-git/go-billy/v5"
-	"github.com/go-git/go-billy/v5/util"
 	"github.com/go-git/go-git/v5"
 	"github.com/go-git/go-git/v5/plumbing"
 	"github.com/go-git/go-git/v5/plumbing/object"
 	"github.com/go-git/go-git/v5/plumbing/transport"
-	"github.com/go-git/go-git/v5/storage/filesystem"
 	"github.com/input-output-hk/catalyst-forge-libs/fs"
-
-	"github.com/input-output-hk/catalyst-forge-libs/git/internal/fsbridge"
 )
 
 const (
@@ -84,23 +77,6 @@ func (o *Options) Validate() error {
 	}
 
 	return nil
-}
-
-// applyDefaults sets default values for any unset fields in Options.
-func (o *Options) applyDefaults() {
-	if o.Workdir == "" {
-		o.Workdir = DefaultWorkdir
-	}
-
-	if o.StorerCacheSize == 0 {
-		o.StorerCacheSize = DefaultStorerCacheSize
-	}
-
-	if o.HTTPClient == nil {
-		o.HTTPClient = &http.Client{
-			Timeout: 30 * time.Second,
-		}
-	}
 }
 
 // AuthProvider resolves authentication methods for git operations.
@@ -179,26 +155,6 @@ type ResolvedRef struct {
 	// CanonicalName is the canonical reference name (e.g., "refs/heads/main").
 	// For commit hashes, this may be the same as Hash.
 	CanonicalName string
-}
-
-// MergeStrategy represents the different types of merge strategies.
-// Currently, only FastForwardOnly is supported by go-git.
-type MergeStrategy int8
-
-const (
-	// FastForwardOnly represents a merge strategy that only allows fast-forward merges.
-	// This will fail if a merge commit would be required.
-	FastForwardOnly MergeStrategy = iota
-)
-
-// String returns a human-readable string representation of the MergeStrategy.
-func (s MergeStrategy) String() string {
-	switch s {
-	case FastForwardOnly:
-		return "fast-forward-only"
-	default:
-		return "unknown"
-	}
 }
 
 // RefKind represents the type of git reference.
@@ -297,11 +253,6 @@ type PatchText struct {
 // Filters are applied progressively - if any filter returns false, the change is excluded.
 type ChangeFilter func(*object.Change) bool
 
-// TagFilter is a predicate function for filtering tags.
-// It returns true if the tag should be included in the results.
-// Filters are applied progressively - if any filter returns false, the tag is excluded.
-type TagFilter func(name string, ref *plumbing.Reference) bool
-
 // Repo represents a git repository and provides high-level operations.
 // It wraps a go-git Repository and Worktree, operating exclusively through
 // the project's native filesystem abstraction.
@@ -310,968 +261,6 @@ type Repo struct {
 	worktree *git.Worktree
 	fs       fs.Filesystem
 	options  Options
-}
-
-// Init creates a new git repository at the specified location.
-// It initializes both bare and non-bare repositories with proper storage and worktree setup.
-func Init(ctx context.Context, opts *Options) (*Repo, error) {
-	if err := opts.Validate(); err != nil {
-		return nil, WrapError(err, "invalid options")
-	}
-
-	opts.applyDefaults()
-
-	// Convert fs.Filesystem to billy.Filesystem
-	billyFS, err := fsbridge.ToBillyFilesystem(opts.FS)
-	if err != nil {
-		return nil, fmt.Errorf("filesystem conversion failed: %w", err)
-	}
-
-	// Chroot to the workdir to scope the repository location
-	scopedFS, err := billyFS.Chroot(opts.Workdir)
-	if err != nil {
-		return nil, fmt.Errorf("failed to chroot to workdir %q: %w", opts.Workdir, err)
-	}
-
-	var storage *filesystem.Storage
-	var worktreeFS gobilly.Filesystem
-
-	if opts.Bare {
-		// For bare repos, storage is at the root
-		storage = fsbridge.NewStorage(scopedFS, opts.StorerCacheSize)
-		worktreeFS = nil
-	} else {
-		// For non-bare repos, storage goes in .git subdirectory
-		dotGitFS, chrootErr := scopedFS.Chroot(".git")
-		if chrootErr != nil {
-			return nil, fmt.Errorf("failed to create .git directory: %w", chrootErr)
-		}
-		storage = fsbridge.NewStorage(dotGitFS, opts.StorerCacheSize)
-		worktreeFS = scopedFS
-	}
-
-	// Initialize repository with custom storage
-	repo, err := git.Init(storage, worktreeFS)
-	if err != nil {
-		return nil, WrapError(err, "failed to initialize repository")
-	}
-
-	r := &Repo{
-		repo:    repo,
-		fs:      opts.FS,
-		options: *opts,
-	}
-
-	// Set up worktree for non-bare repositories
-	if !opts.Bare {
-		worktree, err := repo.Worktree()
-		if err != nil {
-			return nil, WrapError(err, "failed to get worktree")
-		}
-		r.worktree = worktree
-	}
-
-	return r, nil
-}
-
-// CurrentBranch returns the name of the currently checked out branch.
-// It returns an error if HEAD is in a detached state.
-//
-// Context timeout/cancellation is honored during the operation.
-func (r *Repo) CurrentBranch(ctx context.Context) (string, error) {
-	// Get the current HEAD reference
-	head, err := r.repo.Head()
-	if err != nil {
-		return "", WrapError(err, "failed to get HEAD reference")
-	}
-
-	// Check if HEAD is detached (not pointing to a branch)
-	if !head.Name().IsBranch() {
-		return "", WrapError(ErrResolveFailed, "HEAD is detached")
-	}
-
-	// Extract branch name from the reference name
-	branchName := head.Name().Short()
-	return branchName, nil
-}
-
-// CreateBranch creates a new branch from the specified revision.
-// It supports creating branches from any valid revision (commit hash, branch name, tag, etc.).
-// If trackRemote is true, it sets up remote tracking configuration.
-// If force is true, it overwrites any existing branch with the same name.
-//
-// Context timeout/cancellation is honored during the operation.
-func (r *Repo) CreateBranch(ctx context.Context, name, startRev string, trackRemote, force bool) error {
-	if name == "" {
-		return WrapError(ErrInvalidRef, "branch name cannot be empty")
-	}
-
-	if startRev == "" {
-		return WrapError(ErrInvalidRef, "start revision cannot be empty")
-	}
-
-	// Resolve the start revision to a commit hash
-	hash, err := r.repo.ResolveRevision(plumbing.Revision(startRev))
-	if err != nil {
-		return WrapError(ErrResolveFailed, "failed to resolve start revision")
-	}
-
-	// Check if branch already exists
-	branchRefName := plumbing.NewBranchReferenceName(name)
-	_, err = r.repo.Reference(branchRefName, true)
-	if err == nil && !force {
-		return WrapError(ErrBranchExists, "branch already exists")
-	}
-
-	// Create the branch reference
-	newRef := plumbing.NewHashReference(branchRefName, *hash)
-	err = r.repo.Storer.SetReference(newRef)
-	if err != nil {
-		return WrapError(err, "failed to create branch reference")
-	}
-
-	// TODO: Implement remote tracking configuration when trackRemote is true
-	// For now, remote tracking setup is not implemented but the parameter is accepted for API compatibility
-	_ = trackRemote
-
-	return nil
-}
-
-// CheckoutBranch switches to the specified branch.
-// If createIfMissing is true, it creates the branch if it doesn't exist.
-// If force is true, it discards any uncommitted changes in the working tree.
-//
-// Context timeout/cancellation is honored during the operation.
-func (r *Repo) CheckoutBranch(ctx context.Context, name string, createIfMissing, force bool) error {
-	if name == "" {
-		return WrapError(ErrInvalidRef, "branch name cannot be empty")
-	}
-
-	branchRefName := plumbing.NewBranchReferenceName(name)
-
-	// Check if branch exists
-	_, err := r.repo.Reference(branchRefName, true)
-	// Handle non-existent branch
-	if err != nil {
-		if !createIfMissing {
-			return WrapError(ErrBranchMissing, "branch does not exist")
-		}
-
-		// Create the branch from current HEAD
-		head, headErr := r.repo.Head()
-		if headErr != nil {
-			return WrapError(headErr, "failed to get HEAD reference")
-		}
-
-		// Create the branch reference
-		newRef := plumbing.NewHashReference(branchRefName, head.Hash())
-		if setErr := r.repo.Storer.SetReference(newRef); setErr != nil {
-			return WrapError(setErr, "failed to create branch reference")
-		}
-
-		// Verify the branch was created
-		if _, verifyErr := r.repo.Reference(branchRefName, true); verifyErr != nil {
-			// Failed to verify branch creation
-			return WrapError(verifyErr, "branch creation verification failed")
-		}
-	}
-
-	// Now checkout the branch
-	checkoutOpts := &git.CheckoutOptions{
-		Branch: branchRefName,
-		Keep:   false, // Explicitly set keep to false
-	}
-
-	if force {
-		checkoutOpts.Force = true
-	}
-
-	// Store current HEAD reference before checkout
-	currentHead, _ := r.repo.Head()
-
-	err = r.worktree.Checkout(checkoutOpts)
-	if err != nil {
-		return WrapError(err, "failed to checkout branch")
-	}
-
-	// Verify HEAD exists after checkout - if not, restore it
-	if _, headErr := r.repo.Head(); headErr != nil {
-		// HEAD is missing, try to restore it
-		symbolicRef := plumbing.NewSymbolicReference(plumbing.HEAD, branchRefName)
-		if setErr := r.repo.Storer.SetReference(symbolicRef); setErr != nil {
-			// If we can't set symbolic reference, at least try to set a direct reference
-			if currentHead != nil {
-				_ = r.repo.Storer.SetReference(currentHead)
-			}
-			return WrapError(setErr, "failed to restore HEAD after checkout")
-		}
-	}
-
-	return nil
-}
-
-// DeleteBranch deletes the specified local branch.
-// It prevents deletion of the currently checked out branch.
-//
-// Context timeout/cancellation is honored during the operation.
-func (r *Repo) DeleteBranch(ctx context.Context, name string) error {
-	if name == "" {
-		return WrapError(ErrInvalidRef, "branch name cannot be empty")
-	}
-
-	// Get the current branch to prevent deletion of current branch
-	currentBranch, err := r.CurrentBranch(ctx)
-	if err != nil {
-		return WrapError(err, "failed to get current branch")
-	}
-
-	if currentBranch == name {
-		return WrapError(ErrBranchExists, "cannot delete the currently checked out branch")
-	}
-
-	branchRefName := plumbing.NewBranchReferenceName(name)
-
-	// Check if branch exists
-	_, err = r.repo.Reference(branchRefName, true)
-	if err != nil {
-		return WrapError(ErrBranchMissing, "branch does not exist")
-	}
-
-	// Delete the branch
-	err = r.repo.Storer.RemoveReference(branchRefName)
-	if err != nil {
-		return WrapError(err, "failed to delete branch")
-	}
-
-	return nil
-}
-
-// CheckoutRemoteBranch creates a local branch from a remote branch and optionally sets up tracking.
-// If localName is empty, it uses the same name as the remote branch.
-//
-// Context timeout/cancellation is honored during the operation.
-func (r *Repo) CheckoutRemoteBranch(ctx context.Context, remote, remoteBranch, localName string, track bool) error {
-	if remote == "" {
-		return WrapError(ErrInvalidRef, "remote name cannot be empty")
-	}
-
-	if remoteBranch == "" {
-		return WrapError(ErrInvalidRef, "remote branch name cannot be empty")
-	}
-
-	if localName == "" {
-		localName = remoteBranch
-	}
-
-	// Construct the remote branch reference name
-	remoteBranchRef := plumbing.NewRemoteReferenceName(remote, remoteBranch)
-
-	// Check if the remote branch exists
-	remoteRef, err := r.repo.Reference(remoteBranchRef, true)
-	if err != nil {
-		return WrapError(ErrResolveFailed, "remote branch does not exist")
-	}
-
-	// Create the local branch reference
-	localBranchRef := plumbing.NewBranchReferenceName(localName)
-	newRef := plumbing.NewHashReference(localBranchRef, remoteRef.Hash())
-	err = r.repo.Storer.SetReference(newRef)
-	if err != nil {
-		return WrapError(err, "failed to create local branch")
-	}
-
-	// TODO: Implement remote tracking configuration when track is true
-	// For now, remote tracking setup is not implemented but the parameter is accepted for API compatibility
-	_ = track
-
-	// Checkout the newly created local branch
-	checkoutOpts := &git.CheckoutOptions{
-		Branch: localBranchRef,
-	}
-
-	err = r.worktree.Checkout(checkoutOpts)
-	if err != nil {
-		return WrapError(err, "failed to checkout local branch")
-	}
-
-	return nil
-}
-
-// Open discovers and opens an existing git repository at the specified location.
-// It validates that the directory contains a valid git repository structure.
-//
-// The repository must already exist at the specified workdir within the filesystem.
-// For non-bare repositories, both .git directory and worktree must be present.
-// For bare repositories, only the .git directory structure is expected.
-//
-// Context timeout/cancellation is honored during repository validation.
-func Open(ctx context.Context, opts *Options) (*Repo, error) {
-	if err := opts.Validate(); err != nil {
-		return nil, WrapError(err, "invalid options")
-	}
-
-	opts.applyDefaults()
-
-	// Convert fs.Filesystem to billy.Filesystem
-	billyFS, err := fsbridge.ToBillyFilesystem(opts.FS)
-	if err != nil {
-		return nil, fmt.Errorf("filesystem conversion failed: %w", err)
-	}
-
-	// Chroot to the workdir to scope the repository location
-	scopedFS, err := billyFS.Chroot(opts.Workdir)
-	if err != nil {
-		return nil, fmt.Errorf("failed to chroot to workdir %q: %w", opts.Workdir, err)
-	}
-
-	var storage *filesystem.Storage
-	var worktreeFS gobilly.Filesystem
-
-	if opts.Bare {
-		// For bare repos, storage is at the root
-		storage = fsbridge.NewStorage(scopedFS, opts.StorerCacheSize)
-		worktreeFS = nil
-	} else {
-		// For non-bare repos, storage goes in .git subdirectory
-		dotGitFS, chrootErr := scopedFS.Chroot(".git")
-		if chrootErr != nil {
-			return nil, fmt.Errorf("failed to access .git directory: %w", chrootErr)
-		}
-		storage = fsbridge.NewStorage(dotGitFS, opts.StorerCacheSize)
-		worktreeFS = scopedFS
-	}
-
-	// Open existing repository
-	repo, err := git.Open(storage, worktreeFS)
-	if err != nil {
-		return nil, WrapError(err, "failed to open repository")
-	}
-
-	r := &Repo{
-		repo:    repo,
-		fs:      opts.FS,
-		options: *opts,
-	}
-
-	// Set up worktree for non-bare repositories
-	if !opts.Bare {
-		worktree, err := repo.Worktree()
-		if err != nil {
-			return nil, WrapError(err, "failed to get worktree")
-		}
-		r.worktree = worktree
-	}
-
-	return r, nil
-}
-
-// Clone creates a new repository by cloning from a remote URL.
-// It supports both bare and non-bare repositories, shallow cloning, and authentication.
-//
-// The remoteURL should be a valid git URL (https://, ssh://, or file:// for local repos).
-// For shallow clones, set ShallowDepth > 0 to limit the clone depth.
-// Authentication is handled via the AuthProvider if credentials are required.
-//
-// Context timeout/cancellation is honored during the clone operation.
-func Clone(ctx context.Context, remoteURL string, opts *Options) (*Repo, error) {
-	if remoteURL == "" {
-		return nil, WrapError(ErrInvalidRef, "remote URL cannot be empty")
-	}
-
-	if err := opts.Validate(); err != nil {
-		return nil, WrapError(err, "invalid options")
-	}
-
-	opts.applyDefaults()
-
-	// Convert fs.Filesystem to billy.Filesystem
-	billyFS, err := fsbridge.ToBillyFilesystem(opts.FS)
-	if err != nil {
-		return nil, fmt.Errorf("filesystem conversion failed: %w", err)
-	}
-
-	// Chroot to the workdir to scope the repository location
-	scopedFS, err := billyFS.Chroot(opts.Workdir)
-	if err != nil {
-		return nil, fmt.Errorf("failed to chroot to workdir %q: %w", opts.Workdir, err)
-	}
-
-	var storage *filesystem.Storage
-	var worktreeFS gobilly.Filesystem
-
-	if opts.Bare {
-		// For bare repos, storage is at the root
-		storage = fsbridge.NewStorage(scopedFS, opts.StorerCacheSize)
-		worktreeFS = nil
-	} else {
-		// For non-bare repos, storage goes in .git subdirectory
-		dotGitFS, chrootErr := scopedFS.Chroot(".git")
-		if chrootErr != nil {
-			return nil, fmt.Errorf("failed to create .git directory: %w", chrootErr)
-		}
-		storage = fsbridge.NewStorage(dotGitFS, opts.StorerCacheSize)
-		worktreeFS = scopedFS
-	}
-
-	// Prepare clone options
-	cloneOpts := &git.CloneOptions{
-		URL:   remoteURL,
-		Depth: opts.ShallowDepth,
-	}
-
-	// Set up authentication if available
-	if opts.Auth != nil {
-		authMethod, authErr := opts.Auth.Method(remoteURL)
-		if authErr != nil {
-			return nil, WrapError(ErrAuthRequired, "failed to get authentication method")
-		}
-		cloneOpts.Auth = authMethod
-	}
-
-	// Clone the repository
-	repo, err := git.Clone(storage, worktreeFS, cloneOpts)
-	if err != nil {
-		return nil, WrapError(err, "failed to clone repository")
-	}
-
-	r := &Repo{
-		repo:    repo,
-		fs:      opts.FS,
-		options: *opts,
-	}
-
-	// Set up worktree for non-bare repositories
-	if !opts.Bare {
-		worktree, err := repo.Worktree()
-		if err != nil {
-			return nil, WrapError(err, "failed to get worktree")
-		}
-		r.worktree = worktree
-	}
-
-	return r, nil
-}
-
-// Fetch fetches changes from the specified remote.
-// It supports pruning stale remote branches and shallow fetching when depth > 0.
-// Returns ErrAlreadyUpToDate if there are no changes to fetch.
-//
-// Context timeout/cancellation is honored during the fetch operation.
-func (r *Repo) Fetch(ctx context.Context, remote string, prune bool, depth int) error {
-	if remote == "" {
-		remote = DefaultRemoteName
-	}
-
-	// Prepare fetch options
-	fetchOpts := &git.FetchOptions{
-		RemoteName: remote,
-		Prune:      prune,
-		Depth:      depth,
-	}
-
-	// Set up authentication if available
-	if r.options.Auth != nil {
-		// Get the remote URL to determine auth method
-		remoteConfig, err := r.repo.Remote(remote)
-		if err != nil {
-			return WrapError(err, "failed to get remote configuration")
-		}
-
-		authMethod, authErr := r.options.Auth.Method(remoteConfig.Config().URLs[0])
-		if authErr != nil {
-			return WrapError(ErrAuthRequired, "failed to get authentication method")
-		}
-		fetchOpts.Auth = authMethod
-	}
-
-	// Perform the fetch
-	err := r.repo.Fetch(fetchOpts)
-	if err != nil {
-		// Check for specific error types
-		if errors.Is(err, git.ErrRemoteNotFound) {
-			return WrapError(ErrResolveFailed, "remote not found")
-		}
-		if errors.Is(err, git.NoErrAlreadyUpToDate) {
-			return ErrAlreadyUpToDate
-		}
-		return WrapError(err, "failed to fetch from remote")
-	}
-
-	return nil
-}
-
-// PullFFOnly performs a fast-forward only pull from the specified remote.
-// It fetches changes and updates the current branch only if it's a fast-forward merge.
-// Returns ErrNotFastForward if a merge commit would be required.
-// Returns ErrAlreadyUpToDate if there are no changes to pull.
-//
-// Context timeout/cancellation is honored during the pull operation.
-func (r *Repo) PullFFOnly(ctx context.Context, remote string) error {
-	if r.worktree == nil {
-		return WrapError(ErrInvalidRef, "cannot pull in bare repository")
-	}
-
-	if remote == "" {
-		remote = DefaultRemoteName
-	}
-
-	// Prepare pull options with fast-forward only strategy
-	pullOpts := &git.PullOptions{
-		RemoteName: remote,
-	}
-
-	// Set up authentication if available
-	if r.options.Auth != nil {
-		// Get the remote URL to determine auth method
-		remoteConfig, err := r.repo.Remote(remote)
-		if err != nil {
-			return WrapError(err, "failed to get remote configuration")
-		}
-
-		authMethod, authErr := r.options.Auth.Method(remoteConfig.Config().URLs[0])
-		if authErr != nil {
-			return WrapError(ErrAuthRequired, "failed to get authentication method")
-		}
-		pullOpts.Auth = authMethod
-	}
-
-	// Perform the pull
-	err := r.worktree.Pull(pullOpts)
-	if err != nil {
-		// Check for specific error types
-		if errors.Is(err, git.ErrRemoteNotFound) {
-			return WrapError(ErrResolveFailed, "remote not found")
-		}
-		if errors.Is(err, git.NoErrAlreadyUpToDate) {
-			return ErrAlreadyUpToDate
-		}
-		if errors.Is(err, git.ErrNonFastForwardUpdate) {
-			return ErrNotFastForward
-		}
-		return WrapError(err, "failed to pull from remote")
-	}
-
-	return nil
-}
-
-// FetchAndMerge fetches changes from the specified remote and merges the fromRef.
-// It supports different merge strategies as specified by the strategy parameter.
-// Currently, only FastForwardOnly is supported by go-git.
-//
-// Context timeout/cancellation is honored during the operation.
-func (r *Repo) FetchAndMerge(ctx context.Context, remote, fromRef string, strategy MergeStrategy) error {
-	if remote == "" {
-		remote = DefaultRemoteName
-	}
-
-	// First, fetch the changes
-	fetchErr := r.Fetch(ctx, remote, false, 0)
-	if fetchErr != nil && !errors.Is(fetchErr, ErrAlreadyUpToDate) {
-		return WrapError(fetchErr, "failed to fetch before merge")
-	}
-
-	// Resolve the fromRef to a reference
-	hash, err := r.repo.ResolveRevision(plumbing.Revision(fromRef))
-	if err != nil {
-		return WrapError(ErrResolveFailed, "failed to resolve fromRef for merge")
-	}
-
-	ref := plumbing.NewHashReference("", *hash)
-
-	// Prepare merge options
-	var mergeOpts git.MergeOptions
-
-	// Map our strategy to go-git's strategy
-	switch strategy {
-	case FastForwardOnly:
-		mergeOpts.Strategy = git.FastForwardMerge
-	default:
-		return WrapError(ErrInvalidRef, "unsupported merge strategy")
-	}
-
-	// Perform the merge
-	err = r.repo.Merge(*ref, mergeOpts)
-	if err != nil {
-		// Check for specific error types
-		if err.Error() == "unsupported merge strategy" {
-			return WrapError(ErrInvalidRef, "merge strategy not supported by go-git")
-		}
-		// Note: go-git doesn't have a specific error for merge conflicts yet
-		// This may change in future versions
-		return WrapError(err, "failed to merge")
-	}
-
-	return nil
-}
-
-// Remove removes files from the index and worktree.
-// It handles already-deleted files appropriately and supports glob patterns.
-// Files that don't exist in the index are silently ignored.
-//
-// Context timeout/cancellation is honored during the operation.
-func (r *Repo) Remove(ctx context.Context, paths ...string) error {
-	if r.worktree == nil {
-		return WrapError(ErrInvalidRef, "cannot remove files in bare repository")
-	}
-
-	if len(paths) == 0 {
-		return nil // No paths to remove, not an error
-	}
-
-	// Convert fs.Filesystem to billy.Filesystem for glob operations
-	billyFS, err := fsbridge.ToBillyFilesystem(r.fs)
-	if err != nil {
-		return WrapError(err, "failed to convert filesystem for glob operations")
-	}
-
-	// Chroot to workdir for proper path handling
-	workdirFS, err := billyFS.Chroot(r.options.Workdir)
-	if err != nil {
-		return WrapErrorf(err, "failed to chroot to workdir %q", r.options.Workdir)
-	}
-
-	// Collect all paths to remove (expanding globs)
-	var pathsToRemove []string
-
-	for _, path := range paths {
-		if path == "" {
-			continue // Skip empty paths
-		}
-
-		// Check if path contains glob patterns
-		if strings.ContainsAny(path, "*?[") {
-			// Expand glob pattern using billy filesystem
-			matches, globErr := util.Glob(workdirFS, path)
-			if globErr != nil {
-				return WrapErrorf(globErr, "invalid glob pattern %q", path)
-			}
-
-			// Add all matched paths
-			pathsToRemove = append(pathsToRemove, matches...)
-		} else {
-			// Regular path - add it directly
-			pathsToRemove = append(pathsToRemove, path)
-		}
-	}
-
-	// Get current status to check which files are actually tracked
-	currentStatus, err := r.worktree.Status()
-	if err != nil {
-		return WrapError(err, "failed to get worktree status")
-	}
-
-	// Remove each resolved path from the index and worktree
-	for _, path := range pathsToRemove {
-		// Check if file is tracked in the index
-		fileStatus := currentStatus.File(path)
-		if fileStatus.Staging == git.Untracked && fileStatus.Worktree == git.Untracked {
-			// File is not tracked, skip it (matching git rm behavior)
-			continue
-		}
-
-		_, err := r.worktree.Remove(path)
-		if err != nil {
-			return WrapErrorf(err, "failed to remove path %q", path)
-		}
-	}
-
-	return nil
-}
-
-// Unstage unstages files from the index without modifying the worktree.
-// It uses Reset (mixed) to reset the index to HEAD for specified paths.
-// Files that aren't staged are silently ignored.
-//
-// Context timeout/cancellation is honored during the operation.
-func (r *Repo) Unstage(ctx context.Context, paths ...string) error {
-	if r.worktree == nil {
-		return WrapError(ErrInvalidRef, "cannot unstage files in bare repository")
-	}
-
-	if len(paths) == 0 {
-		return nil // No paths to unstage, not an error
-	}
-
-	expandedPaths, err := r.expandPathsForUnstage(paths)
-	if err != nil {
-		return err
-	}
-
-	stagedPaths, err := r.filterStagedPaths(expandedPaths)
-	if err != nil {
-		return err
-	}
-
-	if len(stagedPaths) == 0 {
-		// No staged files to unstage
-		return nil
-	}
-
-	return r.performUnstageReset(stagedPaths)
-}
-
-// expandPathsForUnstage expands glob patterns and filters empty paths
-func (r *Repo) expandPathsForUnstage(paths []string) ([]string, error) {
-	// Convert fs.Filesystem to billy.Filesystem for glob operations
-	billyFS, err := fsbridge.ToBillyFilesystem(r.fs)
-	if err != nil {
-		return nil, WrapError(err, "failed to convert filesystem for glob operations")
-	}
-
-	// Chroot to workdir for proper path handling
-	workdirFS, err := billyFS.Chroot(r.options.Workdir)
-	if err != nil {
-		return nil, WrapErrorf(err, "failed to chroot to workdir %q", r.options.Workdir)
-	}
-
-	// Collect all paths to unstage (expanding globs)
-	var pathsToUnstage []string
-
-	for _, path := range paths {
-		if path == "" {
-			continue // Skip empty paths
-		}
-
-		// Check if path contains glob patterns
-		if strings.ContainsAny(path, "*?[") {
-			// Expand glob pattern using billy filesystem
-			matches, globErr := util.Glob(workdirFS, path)
-			if globErr != nil {
-				return nil, WrapErrorf(globErr, "invalid glob pattern %q", path)
-			}
-
-			// Add all matched paths
-			pathsToUnstage = append(pathsToUnstage, matches...)
-		} else {
-			// Regular path - add it directly
-			pathsToUnstage = append(pathsToUnstage, path)
-		}
-	}
-
-	return pathsToUnstage, nil
-}
-
-// filterStagedPaths returns only the paths that are actually staged
-func (r *Repo) filterStagedPaths(paths []string) ([]string, error) {
-	// Get current status to check which files are actually staged
-	currentStatus, statusErr := r.worktree.Status()
-	if statusErr != nil {
-		return nil, WrapError(statusErr, "failed to get worktree status")
-	}
-
-	// Collect only the paths that are actually staged
-	var stagedPaths []string
-	for _, path := range paths {
-		fileStatus := currentStatus.File(path)
-		if fileStatus.Staging != git.Untracked && fileStatus.Staging != git.Unmodified {
-			// File is staged (added, modified, deleted, etc.), so we can unstage it
-			stagedPaths = append(stagedPaths, path)
-		}
-		// Silently ignore files that aren't staged (matching git reset behavior)
-	}
-
-	return stagedPaths, nil
-}
-
-// performUnstageReset performs the actual reset operation to unstage files
-func (r *Repo) performUnstageReset(stagedPaths []string) error {
-	// Get HEAD commit to reset to
-	head, err := r.repo.Head()
-	if err != nil {
-		return WrapError(err, "failed to get HEAD reference")
-	}
-
-	// Use Reset to unstage specific paths by resetting index to HEAD
-	resetOpts := &git.ResetOptions{
-		Commit: head.Hash(),
-		Mode:   git.MixedReset,
-		Files:  stagedPaths,
-	}
-	err = r.worktree.Reset(resetOpts)
-	if err != nil {
-		return WrapError(err, "failed to unstage files")
-	}
-
-	return nil
-}
-
-// Commit creates a new commit with the specified message and author/committer.
-// It returns the SHA of the new commit. The CommitOpts can be used to control
-// commit behavior such as allowing empty commits.
-//
-// Context timeout/cancellation is honored during the operation.
-func (r *Repo) Commit(ctx context.Context, msg string, who Signature, opts CommitOpts) (string, error) {
-	if r.worktree == nil {
-		return "", WrapError(ErrInvalidRef, "cannot commit in bare repository")
-	}
-
-	if msg == "" {
-		return "", WrapError(ErrInvalidRef, "commit message cannot be empty")
-	}
-
-	if who.Name == "" || who.Email == "" {
-		return "", WrapError(ErrInvalidRef, "committer name and email are required")
-	}
-
-	// Check if there are any staged changes
-	status, err := r.worktree.Status()
-	if err != nil {
-		return "", WrapError(err, "failed to get worktree status")
-	}
-
-	// Count staged files
-	stagedCount := 0
-	for _, fileStatus := range status {
-		if fileStatus.Staging != git.Untracked && fileStatus.Staging != git.Unmodified {
-			stagedCount++
-		}
-	}
-
-	if stagedCount == 0 && !opts.AllowEmpty {
-		return "", WrapError(ErrEmptyCommit, "no changes staged for commit")
-	}
-
-	// Prepare commit options
-	commitOpts := &git.CommitOptions{
-		Author: &object.Signature{
-			Name:  who.Name,
-			Email: who.Email,
-			When:  who.When,
-		},
-		Committer: &object.Signature{
-			Name:  who.Name,
-			Email: who.Email,
-			When:  who.When,
-		},
-		AllowEmptyCommits: opts.AllowEmpty,
-	}
-
-	// Create the commit
-	hash, err := r.worktree.Commit(msg, commitOpts)
-	if err != nil {
-		if errors.Is(err, git.ErrEmptyCommit) {
-			return "", ErrEmptyCommit
-		}
-		return "", WrapError(err, "failed to create commit")
-	}
-
-	return hash.String(), nil
-}
-
-// Push pushes the current branch to the specified remote.
-// It supports force pushing when force is true.
-// Returns ErrNotFastForward if the push would overwrite remote changes and force is false.
-// Returns ErrAlreadyUpToDate if there are no changes to push.
-//
-// Context timeout/cancellation is honored during the push operation.
-func (r *Repo) Push(ctx context.Context, remote string, force bool) error {
-	if remote == "" {
-		remote = DefaultRemoteName
-	}
-
-	// Prepare push options
-	pushOpts := &git.PushOptions{
-		RemoteName: remote,
-		Force:      force,
-	}
-
-	// Set up authentication if available
-	if r.options.Auth != nil {
-		// Get the remote URL to determine auth method
-		remoteConfig, err := r.repo.Remote(remote)
-		if err != nil {
-			return WrapError(err, "failed to get remote configuration")
-		}
-
-		authMethod, authErr := r.options.Auth.Method(remoteConfig.Config().URLs[0])
-		if authErr != nil {
-			return WrapError(ErrAuthRequired, "failed to get authentication method")
-		}
-		pushOpts.Auth = authMethod
-	}
-
-	// Perform the push
-	err := r.repo.Push(pushOpts)
-	if err != nil {
-		// Check for specific error types
-		if errors.Is(err, git.ErrRemoteNotFound) {
-			return WrapError(ErrResolveFailed, "remote not found")
-		}
-		if errors.Is(err, git.NoErrAlreadyUpToDate) {
-			return ErrAlreadyUpToDate
-		}
-		if errors.Is(err, git.ErrNonFastForwardUpdate) {
-			return ErrNotFastForward
-		}
-		return WrapError(err, "failed to push to remote")
-	}
-
-	return nil
-}
-
-// Add stages files in the worktree for the next commit.
-// It supports glob patterns and handles missing files appropriately.
-// Files that don't exist are silently ignored (matching git add behavior).
-//
-// Context timeout/cancellation is honored during the operation.
-func (r *Repo) Add(ctx context.Context, paths ...string) error {
-	if r.worktree == nil {
-		return WrapError(ErrInvalidRef, "cannot add files in bare repository")
-	}
-
-	if len(paths) == 0 {
-		return nil // No paths to add, not an error
-	}
-
-	// Convert fs.Filesystem to billy.Filesystem for glob operations
-	billyFS, err := fsbridge.ToBillyFilesystem(r.fs)
-	if err != nil {
-		return WrapError(err, "failed to convert filesystem for glob operations")
-	}
-
-	// Chroot to workdir for proper path handling
-	workdirFS, err := billyFS.Chroot(r.options.Workdir)
-	if err != nil {
-		return WrapErrorf(err, "failed to chroot to workdir %q", r.options.Workdir)
-	}
-
-	// Collect all paths to add (expanding globs and filtering out non-existent files)
-	var pathsToAdd []string
-
-	for _, path := range paths {
-		if path == "" {
-			continue // Skip empty paths
-		}
-
-		// Check if path contains glob patterns
-		if strings.ContainsAny(path, "*?[") {
-			// Expand glob pattern using billy filesystem
-			matches, globErr := util.Glob(workdirFS, path)
-			if globErr != nil {
-				return WrapErrorf(globErr, "invalid glob pattern %q", path)
-			}
-
-			// Add all matched paths
-			pathsToAdd = append(pathsToAdd, matches...)
-		} else {
-			// Regular path - check if it exists
-			exists, err := workdirFS.Stat(path)
-			if err == nil && exists != nil {
-				pathsToAdd = append(pathsToAdd, path)
-			}
-			// Silently ignore non-existent files (matching git behavior)
-		}
-	}
-
-	// Add each resolved path to the index
-	for _, path := range pathsToAdd {
-		_, err := r.worktree.Add(path)
-		if err != nil {
-			return WrapErrorf(err, "failed to add path %q", path)
-		}
-	}
-
-	return nil
 }
 
 // Log returns a commit iterator for the repository with the specified filters applied.
@@ -1630,236 +619,6 @@ func ChangeExcludePathFilter(path string) ChangeFilter {
 	includeFilter := ChangePathFilter(path)
 	return func(change *object.Change) bool {
 		return !includeFilter(change) // Invert the path filter
-	}
-}
-
-// CreateTag creates a new tag at the specified target revision.
-// If message is provided and annotated is true, an annotated tag is created.
-// If message is empty or annotated is false, a lightweight tag is created.
-// The target can be any valid revision specifier (commit hash, branch name, etc.).
-//
-// Context timeout/cancellation is honored during the operation.
-func (r *Repo) CreateTag(ctx context.Context, name, target, message string, annotated bool) error {
-	if name == "" {
-		return WrapError(ErrInvalidRef, "tag name cannot be empty")
-	}
-
-	if target == "" {
-		return WrapError(ErrInvalidRef, "target revision cannot be empty")
-	}
-
-	// Resolve the target revision to a commit hash
-	hash, err := r.repo.ResolveRevision(plumbing.Revision(target))
-	if err != nil {
-		return WrapError(ErrResolveFailed, "failed to resolve target revision")
-	}
-
-	// Check if tag already exists
-	tagRefName := plumbing.NewTagReferenceName(name)
-	_, err = r.repo.Reference(tagRefName, true)
-	if err == nil {
-		return WrapError(ErrTagExists, "tag already exists")
-	}
-
-	// Create the tag based on type
-	if annotated && message != "" {
-		// Create annotated tag
-		tagOpts := &git.CreateTagOptions{
-			Tagger: &object.Signature{
-				Name:  "git-wrapper", // Default tagger - could be made configurable
-				Email: "git@catalyst-forge-libs",
-				When:  time.Now(),
-			},
-			Message: message,
-		}
-
-		_, err = r.repo.CreateTag(name, *hash, tagOpts)
-		if err != nil {
-			return WrapError(err, "failed to create annotated tag")
-		}
-	} else {
-		// Create lightweight tag
-		tagRef := plumbing.NewHashReference(tagRefName, *hash)
-		err = r.repo.Storer.SetReference(tagRef)
-		if err != nil {
-			return WrapError(err, "failed to create lightweight tag")
-		}
-	}
-
-	return nil
-}
-
-// DeleteTag deletes the specified tag from the repository.
-// Returns ErrTagMissing if the tag does not exist.
-//
-// Context timeout/cancellation is honored during the operation.
-func (r *Repo) DeleteTag(ctx context.Context, name string) error {
-	if name == "" {
-		return WrapError(ErrInvalidRef, "tag name cannot be empty")
-	}
-
-	// Check if tag exists
-	tagRefName := plumbing.NewTagReferenceName(name)
-	_, err := r.repo.Reference(tagRefName, true)
-	if err != nil {
-		return WrapError(ErrTagMissing, "tag does not exist")
-	}
-
-	// Delete the tag
-	err = r.repo.Storer.RemoveReference(tagRefName)
-	if err != nil {
-		return WrapError(err, "failed to delete tag")
-	}
-
-	return nil
-}
-
-// Tags returns a list of tags that pass all the provided filters.
-// If no filters are provided, all tags are returned.
-// Filters are applied progressively - a tag must pass ALL filters to be included.
-// Results are sorted alphabetically.
-//
-// Context timeout/cancellation is honored during the operation.
-func (r *Repo) Tags(ctx context.Context, filters ...TagFilter) ([]string, error) {
-	// Get all tag references
-	refs, err := r.repo.References()
-	if err != nil {
-		return nil, WrapError(err, "failed to get references")
-	}
-
-	var tags []string
-	err = refs.ForEach(func(ref *plumbing.Reference) error {
-		if ref.Name().IsTag() {
-			tagName := ref.Name().Short()
-
-			// Apply filters
-			if shouldIncludeTag(tagName, ref, filters) {
-				tags = append(tags, tagName)
-			}
-		}
-		return nil
-	})
-	if err != nil {
-		return nil, WrapError(err, "failed to iterate references")
-	}
-
-	// Sort tags alphabetically
-	for i := 0; i < len(tags)-1; i++ {
-		for j := i + 1; j < len(tags); j++ {
-			if tags[i] > tags[j] {
-				tags[i], tags[j] = tags[j], tags[i]
-			}
-		}
-	}
-
-	return tags, nil
-}
-
-// shouldIncludeTag checks if a tag passes all filters
-func shouldIncludeTag(name string, ref *plumbing.Reference, filters []TagFilter) bool {
-	for _, filter := range filters {
-		if filter != nil && !filter(name, ref) {
-			return false
-		}
-	}
-	return true
-}
-
-// Common TagFilter implementations for convenience
-
-// TagPatternFilter returns a filter that matches tags against a glob pattern.
-// Supports * (matches any number of characters) and ? (matches single character).
-// For example: "v1.*" matches "v1.0", "v1.1", etc.
-func TagPatternFilter(pattern string) TagFilter {
-	return func(name string, ref *plumbing.Reference) bool {
-		return matchesTagPattern(name, pattern)
-	}
-}
-
-// matchesTagPattern checks if a tag name matches the given pattern
-func matchesTagPattern(name, pattern string) bool {
-	if pattern == "" {
-		return true // Empty pattern matches all
-	}
-
-	// Handle * wildcard
-	if strings.Contains(pattern, "*") {
-		return matchesStarPattern(name, pattern)
-	}
-
-	// Handle ? wildcard
-	if strings.Contains(pattern, "?") {
-		return matchesQuestionPattern(name, pattern)
-	}
-
-	// Exact match for patterns without wildcards
-	return name == pattern
-}
-
-// matchesStarPattern handles * wildcard matching
-func matchesStarPattern(name, pattern string) bool {
-	parts := strings.Split(pattern, "*")
-	remaining := name
-
-	for i, part := range parts {
-		if part == "" {
-			continue // Skip empty parts
-		}
-
-		idx := strings.Index(remaining, part)
-		if idx < 0 {
-			return false // Part not found
-		}
-
-		// First part must be at beginning unless pattern starts with *
-		if i == 0 && pattern[0] != '*' && idx != 0 {
-			return false
-		}
-
-		remaining = remaining[idx+len(part):]
-	}
-
-	// Check if pattern ends with * or remaining is empty
-	return pattern[len(pattern)-1] == '*' || remaining == ""
-}
-
-// matchesQuestionPattern handles ? wildcard matching
-func matchesQuestionPattern(name, pattern string) bool {
-	if len(pattern) != len(name) {
-		return false // Different lengths can't match
-	}
-
-	for i := 0; i < len(pattern); i++ {
-		if pattern[i] != '?' && pattern[i] != name[i] {
-			return false
-		}
-	}
-	return true
-}
-
-// TagPrefixFilter returns a filter that matches tags with the given prefix.
-// For example: "v" matches "v1.0", "v2.0", etc.
-func TagPrefixFilter(prefix string) TagFilter {
-	return func(name string, ref *plumbing.Reference) bool {
-		return strings.HasPrefix(name, prefix)
-	}
-}
-
-// TagSuffixFilter returns a filter that matches tags with the given suffix.
-// For example: "-rc" matches "v1.0-rc", "v2.0-rc", etc.
-func TagSuffixFilter(suffix string) TagFilter {
-	return func(name string, ref *plumbing.Reference) bool {
-		return strings.HasSuffix(name, suffix)
-	}
-}
-
-// TagExcludeFilter returns a filter that excludes tags matching the given pattern.
-// This is useful for filtering out certain tags while keeping all others.
-// For example: TagExcludeFilter("*-rc") excludes all release candidates.
-func TagExcludeFilter(pattern string) TagFilter {
-	includeFilter := TagPatternFilter(pattern)
-	return func(name string, ref *plumbing.Reference) bool {
-		return !includeFilter(name, ref) // Invert the pattern filter
 	}
 }
