+++ git/lifecycle.go
@@ -1,0 +1,250 @@
+// Package git provides lifecycle operations for git repositories.
+// This file contains the core repository initialization, opening, and cloning operations.
+package git
+
+import (
+	"context"
+	"fmt"
+	"net/http"
+	"time"
+
+	gobilly "github.com/go-git/go-billy/v5"
+	"github.com/go-git/go-git/v5"
+	"github.com/go-git/go-git/v5/storage/filesystem"
+
+	"github.com/input-output-hk/catalyst-forge-libs/git/internal/fsbridge"
+)
+
+// applyDefaults sets default values for any unset fields in Options.
+func (o *Options) applyDefaults() {
+	if o.Workdir == "" {
+		o.Workdir = DefaultWorkdir
+	}
+
+	if o.StorerCacheSize == 0 {
+		o.StorerCacheSize = DefaultStorerCacheSize
+	}
+
+	if o.HTTPClient == nil {
+		o.HTTPClient = &http.Client{
+			Timeout: 30 * time.Second,
+		}
+	}
+}
+
+// Init creates a new git repository at the specified location.
+// It initializes both bare and non-bare repositories with proper storage and worktree setup.
+func Init(ctx context.Context, opts *Options) (*Repo, error) {
+	if err := opts.Validate(); err != nil {
+		return nil, WrapError(err, "invalid options")
+	}
+
+	opts.applyDefaults()
+
+	// Convert fs.Filesystem to billy.Filesystem
+	billyFS, err := fsbridge.ToBillyFilesystem(opts.FS)
+	if err != nil {
+		return nil, fmt.Errorf("filesystem conversion failed: %w", err)
+	}
+
+	// Chroot to the workdir to scope the repository location
+	scopedFS, err := billyFS.Chroot(opts.Workdir)
+	if err != nil {
+		return nil, fmt.Errorf("failed to chroot to workdir %q: %w", opts.Workdir, err)
+	}
+
+	var storage *filesystem.Storage
+	var worktreeFS gobilly.Filesystem
+
+	if opts.Bare {
+		// For bare repos, storage is at the root
+		storage = fsbridge.NewStorage(scopedFS, opts.StorerCacheSize)
+		worktreeFS = nil
+	} else {
+		// For non-bare repos, storage goes in .git subdirectory
+		dotGitFS, chrootErr := scopedFS.Chroot(".git")
+		if chrootErr != nil {
+			return nil, fmt.Errorf("failed to create .git directory: %w", chrootErr)
+		}
+		storage = fsbridge.NewStorage(dotGitFS, opts.StorerCacheSize)
+		worktreeFS = scopedFS
+	}
+
+	// Initialize repository with custom storage
+	repo, err := git.Init(storage, worktreeFS)
+	if err != nil {
+		return nil, WrapError(err, "failed to initialize repository")
+	}
+
+	r := &Repo{
+		repo:    repo,
+		fs:      opts.FS,
+		options: *opts,
+	}
+
+	// Set up worktree for non-bare repositories
+	if !opts.Bare {
+		worktree, err := repo.Worktree()
+		if err != nil {
+			return nil, WrapError(err, "failed to get worktree")
+		}
+		r.worktree = worktree
+	}
+
+	return r, nil
+}
+
+// Open discovers and opens an existing git repository at the specified location.
+// It validates that the directory contains a valid git repository structure.
+//
+// The repository must already exist at the specified workdir within the filesystem.
+// For non-bare repositories, both .git directory and worktree must be present.
+// For bare repositories, only the .git directory structure is expected.
+//
+// Context timeout/cancellation is honored during repository validation.
+func Open(ctx context.Context, opts *Options) (*Repo, error) {
+	if err := opts.Validate(); err != nil {
+		return nil, WrapError(err, "invalid options")
+	}
+
+	opts.applyDefaults()
+
+	// Convert fs.Filesystem to billy.Filesystem
+	billyFS, err := fsbridge.ToBillyFilesystem(opts.FS)
+	if err != nil {
+		return nil, fmt.Errorf("filesystem conversion failed: %w", err)
+	}
+
+	// Chroot to the workdir to scope the repository location
+	scopedFS, err := billyFS.Chroot(opts.Workdir)
+	if err != nil {
+		return nil, fmt.Errorf("failed to chroot to workdir %q: %w", opts.Workdir, err)
+	}
+
+	var storage *filesystem.Storage
+	var worktreeFS gobilly.Filesystem
+
+	if opts.Bare {
+		// For bare repos, storage is at the root
+		storage = fsbridge.NewStorage(scopedFS, opts.StorerCacheSize)
+		worktreeFS = nil
+	} else {
+		// For non-bare repos, storage goes in .git subdirectory
+		dotGitFS, chrootErr := scopedFS.Chroot(".git")
+		if chrootErr != nil {
+			return nil, fmt.Errorf("failed to access .git directory: %w", chrootErr)
+		}
+		storage = fsbridge.NewStorage(dotGitFS, opts.StorerCacheSize)
+		worktreeFS = scopedFS
+	}
+
+	// Open existing repository
+	repo, err := git.Open(storage, worktreeFS)
+	if err != nil {
+		return nil, WrapError(err, "failed to open repository")
+	}
+
+	r := &Repo{
+		repo:    repo,
+		fs:      opts.FS,
+		options: *opts,
+	}
+
+	// Set up worktree for non-bare repositories
+	if !opts.Bare {
+		worktree, err := repo.Worktree()
+		if err != nil {
+			return nil, WrapError(err, "failed to get worktree")
+		}
+		r.worktree = worktree
+	}
+
+	return r, nil
+}
+
+// Clone creates a new repository by cloning from a remote URL.
+// It supports both bare and non-bare repositories, shallow cloning, and authentication.
+//
+// The remoteURL should be a valid git URL (https://, ssh://, or file:// for local repos).
+// For shallow clones, set ShallowDepth > 0 to limit the clone depth.
+// Authentication is handled via the AuthProvider if credentials are required.
+//
+// Context timeout/cancellation is honored during the clone operation.
+func Clone(ctx context.Context, remoteURL string, opts *Options) (*Repo, error) {
+	if remoteURL == "" {
+		return nil, WrapError(ErrInvalidRef, "remote URL cannot be empty")
+	}
+
+	if err := opts.Validate(); err != nil {
+		return nil, WrapError(err, "invalid options")
+	}
+
+	opts.applyDefaults()
+
+	// Convert fs.Filesystem to billy.Filesystem
+	billyFS, err := fsbridge.ToBillyFilesystem(opts.FS)
+	if err != nil {
+		return nil, fmt.Errorf("filesystem conversion failed: %w", err)
+	}
+
+	// Chroot to the workdir to scope the repository location
+	scopedFS, err := billyFS.Chroot(opts.Workdir)
+	if err != nil {
+		return nil, fmt.Errorf("failed to chroot to workdir %q: %w", opts.Workdir, err)
+	}
+
+	var storage *filesystem.Storage
+	var worktreeFS gobilly.Filesystem
+
+	if opts.Bare {
+		// For bare repos, storage is at the root
+		storage = fsbridge.NewStorage(scopedFS, opts.StorerCacheSize)
+		worktreeFS = nil
+	} else {
+		// For non-bare repos, storage goes in .git subdirectory
+		dotGitFS, chrootErr := scopedFS.Chroot(".git")
+		if chrootErr != nil {
+			return nil, fmt.Errorf("failed to create .git directory: %w", chrootErr)
+		}
+		storage = fsbridge.NewStorage(dotGitFS, opts.StorerCacheSize)
+		worktreeFS = scopedFS
+	}
+
+	// Prepare clone options
+	cloneOpts := &git.CloneOptions{
+		URL:   remoteURL,
+		Depth: opts.ShallowDepth,
+	}
+
+	// Set up authentication if available
+	if opts.Auth != nil {
+		authMethod, authErr := opts.Auth.Method(remoteURL)
+		if authErr != nil {
+			return nil, WrapError(ErrAuthRequired, "failed to get authentication method")
+		}
+		cloneOpts.Auth = authMethod
+	}
+
+	// Clone the repository
+	repo, err := git.Clone(storage, worktreeFS, cloneOpts)
+	if err != nil {
+		return nil, WrapError(err, "failed to clone repository")
+	}
+
+	r := &Repo{
+		repo:    repo,
+		fs:      opts.FS,
+		options: *opts,
+	}
+
+	// Set up worktree for non-bare repositories
+	if !opts.Bare {
+		worktree, err := repo.Worktree()
+		if err != nil {
+			return nil, WrapError(err, "failed to get worktree")
+		}
+		r.worktree = worktree
+	}
+
+	return r, nil
+}
